chore: move ModlogFeed to domain model
