+++ lib/src/features/modlog/modlog.dart
@@ -3,4 +3,5 @@ export 'presentation/pages/pages.dart';
 export 'presentation/state/modlog_cubit.dart';
 export 'presentation/widgets/widgets.dart';
 export 'domain/enums/enums.dart';
+export 'domain/models/modlog_feed.dart';
 export 'data/repositories/modlog_repository.dart';
